refactor: restructuring as per jina docs
