+++ config.yml
@@ -3,7 +3,4 @@ with:
   {}
 metas:
   py_modules:
-    - ast_exec/ast_input.py
-    - ast_exec/ast_models.py
-    - ast_exec/ast_params.py
-    - ast_encoder.py
+    - executor/__init__.py
